Controller specs. Rcov to 100%. Refactored helper.
Got to an interesting problem, worth sharing. I was doing it wrong by
calling a model in the helper, and it showed when I attempted to test
it - so it got refactored. This would also cause problems if you start
profiling the code (showing exsessive rendering times, hiding blasphemy
in the helpers).

I also got back a bit, to spec the controller create action (my mistake
for not doing it when I had to).